move GAW reader to own plugin submodule
