+++ pyaerocom/io/__init__.py
@@ -69,7 +69,6 @@ from .read_earlinet import ReadEarlinet
 from .read_ebas import ReadEbas
 from .read_eea_aqerep import ReadEEAAQEREP
 from .read_eea_aqerep_v2 import ReadEEAAQEREP_V2
-from .read_gaw import ReadGAW
 from .read_ghost import ReadGhost
 from .read_marcopolo import ReadMarcoPolo
