+++ pyaerocom/io/readungridded.py
@@ -20,9 +20,9 @@ from pyaerocom.io.read_earlinet import ReadEarlinet
 from pyaerocom.io.read_ebas import ReadEbas
 from pyaerocom.io.read_eea_aqerep import ReadEEAAQEREP
 from pyaerocom.io.read_eea_aqerep_v2 import ReadEEAAQEREP_V2
-from pyaerocom.io.read_gaw import ReadGAW
 from pyaerocom.io.read_ghost import ReadGhost
 from pyaerocom.io.read_marcopolo import ReadMarcoPolo
+from pyaerocom.plugins.gaw.reader import ReadGAW
 from pyaerocom.ungriddeddata import UngriddedData
 from pyaerocom.variable import get_aliases
